% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{getPvals}
\alias{getPvals}
\title{Estimate the empirical p-values}
\usage{
getPvals(permuteResult, scoredf)
}
\arguments{
\item{permuteResult}{A matrix, null distributions for each sample generated
using the \code{\link[=generateNull]{generateNull()}} function}

\item{scoredf}{A dataframe, singscores generated using the \code{\link[=simpleScore]{simpleScore()}}
function}
}
\value{
Estimated p-values for enrichment of the gene signature in each
sample. A p-value of 1/B indicates that the estimated p-value is less than
or equal to 1/B.
}
\description{
This function takes the null distribution obtained by using
\code{\link[=generateNull]{generateNull()}}, and the calculated 'singscore' using \code{\link[=simpleScore]{simpleScore()}} to
estimate the empirical p-values. The p-values are estimated using a
two-tailed test. A minimum p-value of 1/B can be achieved with B
permutations. using formula p = (r+1)/(m+1) where r is the number of
empirical scores that are larger than the obtained score and m is the total
number of permutation run which is the B parameter in \code{\link[=generateNull]{generateNull()}}
}
\examples{
ranked <- rankGenes(toy_expr)
scoredf <- simpleScore(ranked, upSet = toy_gs_up, downSet = toy_gs_dn)
n_up = length(GSEABase::geneIds(toy_gs_up))
n_down = length(GSEABase::geneIds(toy_gs_dn))
# find out what backends can be registered on your machine
BiocParallel::registered()
# the first one is the default backend, and it can be changed explicitly.
permuteResult = generateNull(n_up = n_up, n_down = n_down, ranked, B =10,
seed = 1) # call the permutation function to generate the empirical scores 
#for B times.
pvals <- getPvals(permuteResult,scoredf)
}
\author{
Ruqian Lyu
}
