% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{permute_null}
\alias{permute_null}
\title{Permutation test for the derived scores of each sample}
\usage{
permute_null(n_up, n_down, rankData, B = 1000, seed = 1)
}
\arguments{
\item{n_up}{integer,  size of up set}

\item{n_down}{integer, size of down set}

\item{rankData}{matrix, outcome of function \code{\link[=rankExpr]{rankExpr()}}}

\item{B}{integer, the number of permutation repeats default as 1000}

\item{seed}{integer, set the seed for randomisation}
}
\value{
A matrix of empircal scores for each sample
}
\description{
This function randomly generates a number of gene sets which
have the same number of genes as the scored gene set. It scores each random
gene set and return a matrix with scores for each permutation across all
samples. The empirical scores are used to calculate the empirical p value
and plot the null distribution. The implementation uses
\code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}} for easy access to parallel backends.
}
\examples{
ranked <- rankExpr(toy_expr)
scoredf <- singscoring(ranked, upSet = toy_up, downSet = toy_dn)
n_up = length(GSEABase::geneIds(toy_up))
n_down = length(GSEABase::geneIds(toy_dn))
# find out what backends can be registered on your machine
BiocParallel::registered()
# the first one is the default backend, can be changed explicitly.

permuteResult = permute_null(n_up = n_up, n_down = n_down, ranked, B =10,
seed = 1) # call the permutation function to generate the empirical scores 
#for B times.
}
\seealso{
\link{Post about BiocParallel}
{http://lcolladotor.github.io/2016/03/07/BiocParallel/#.WgXMF61L28U}
\code{browseVignettes("BiocParallel")}
}
