% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{plotRankDensity_intl}
\alias{plotRankDensity_intl}
\title{Plot the densities of ranks for one sample}
\usage{
plotRankDensity_intl(rankData, upSet, downSet = NULL, isInteractive = FALSE,
  textSize = 1.5)
}
\arguments{
\item{rankData}{one column of the ranked gene expression matrix obtained from
the \code{\link[=rankExpr]{rankExpr()}} function, use drop = FALSE when subsetting the ranked gene
expression matrix, see examples.}

\item{upSet}{GeneSet object, up regulated gene set}

\item{downSet}{GeneSet object, down regulated gene set}

\item{isInteractive}{Boolean, determin whether the returned plot is
interactive}

\item{textSize}{numberic, set the size of text on the plot}
}
\value{
A ggplot object (optionally interactive) demonstrating the rank
density along with rug plot
}
\description{
This function takes a single column data frame, which is a
subset of the ranked data obtained from \code{\link[=rankExpr]{rankExpr()}}function, and gene sets,
and it returns plots visualising the density and the rugs of the ran ks.
}
\seealso{
\code{"\linkS4class{GeneSet}"}
}
