% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{plotScoreLandscape}
\alias{plotScoreLandscape}
\title{Plot landscape of two gene signatures scores}
\usage{
plotScoreLandscape(scoredf1, scoredf2, scorenames = c(), textSize = 1.5,
  isInteractive = FALSE, hexMin = 100)
}
\arguments{
\item{scoredf1}{data.frame, result of the singscoring() function which scores
the gene expression matrix against a gene set of interest}

\item{scoredf2}{data.frame, result of the singscoring() function which scores
the gene expression matrix against another gene set of interest}

\item{scorenames}{character vector of length 2, names for the two scored gene
set/signatures stored in scoredf1 and scoredf2}

\item{textSize}{numeric, set the text size for the plot, default as 1.5}

\item{isInteractive}{boolean, whether the plot is interactive default as
FALSE}

\item{hexMin}{integer, the threshold which decides whether hex bin plot or
scatter plot is displayed, default as 100}
}
\value{
A ggplot object, a scatter plot, demostrating the relationship
between scores from two signatures on the same set of samples.
}
\description{
This function takes two data frames which are outputs from the
singscoring() function and plots the relationship between the two gene set
scores for samples in the gene expression matrix.
}
\examples{
ranked <- rankExpr(toy_expr)
scoredf <- singscoring(ranked, upSet = toy_up, downSet = toy_dn)
scoredf2 <- singscoring(ranked, upSet = toy_up)
plotScoreLandscape(scoredf, scoredf2)
}
