% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{projectScoreLandscape}
\alias{projectScoreLandscape}
\title{Project the landscape score obtained from plotScoreLandscape}
\usage{
projectScoreLandscape(plotObj = NULL, scoredf1, scoredf2,
  sampleLabels = NULL, annot = NULL, labels = 20, isInteractive = F)
}
\arguments{
\item{plotObj}{a dataframe, resulted from plotScoreLanscape()}

\item{scoredf1}{data.frame, results of the simpleScore() or singscoring() function}

\item{scoredf2}{data.frame, results of the simpleScore() or singscoring() function}

\item{sampleLabels}{character, sample names to display, ordered in the same way
as samples are ordered in the 'singscored' data matrix}

\item{annot}{Annotations to colour the data.}

\item{labels}{integer, number of samples labeled on the plot in the new data}

\item{isInteractive}{boolean, whether the plot is interactive}
}
\value{
New data points on the already plotted ggplot object from
plotScoreLanscape()
}
\description{
This function takes the output of the plotScoreLandscape() and
a new data, and it projects the new data onto the ggplot object.
}
\examples{
ranked <- rankGenes(toy_expr)
scoredf <- simpleScore(ranked, upSet = toy_up, downSet = toy_dn)
scoredf2 <- simpleScore(ranked, upSet = toy_up)
psl <- plotScoreLandscape(scoredf, scoredf2)
projectScoreLandscape(psl,scoredf, scoredf2)
}
