% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankExprGeneric.R
\docType{methods}
\name{rankExpr}
\alias{rankExpr}
\alias{rankExpr,matrix,ANY-method}
\alias{rankExpr,matrix,character-method}
\alias{rankExpr,data.frame,ANY-method}
\alias{rankExpr,data.frame,character-method}
\alias{rankExpr,DGEList,character-method}
\alias{rankExpr,DGEList,ANY-method}
\alias{rankExpr,ExpressionSet,ANY-method}
\alias{rankExpr,ExpressionSet,character-method}
\title{Rank genes by the gene expression intensities}
\usage{
rankExpr(expreMatrix, tiesMethod)

\S4method{rankExpr}{matrix,ANY}(expreMatrix)

\S4method{rankExpr}{matrix,character}(expreMatrix, tiesMethod)

\S4method{rankExpr}{data.frame,ANY}(expreMatrix)

\S4method{rankExpr}{data.frame,character}(expreMatrix, tiesMethod)

\S4method{rankExpr}{DGEList,character}(expreMatrix, tiesMethod)

\S4method{rankExpr}{DGEList,ANY}(expreMatrix)

\S4method{rankExpr}{ExpressionSet,ANY}(expreMatrix)

\S4method{rankExpr}{ExpressionSet,character}(expreMatrix, tiesMethod)
}
\arguments{
\item{expreMatrix}{A gene expression matrix or S4 object
(ExpressionSet,DGEList,data.frame)}

\item{tiesMethod}{A character indicating what method to use when dealing with
ties}
}
\value{
The ranked gene expression matrix that has samples in columns and
 genes in rows
}
\description{
The \code{rankExpr} function is a generic function that takes
mutilple combinations of inputs.If input is S4 object of \code{DGEList, or
ExpressionSet}, the generic function will extract the gene epxression matrix
genes on the rows and samples on columns from the object and rank the genes.
Given a matrix of gene expression that has samples in columns, genes in rows,
and values being gene expression intensity. \code{rankExpr} ranks gene
expression in each sample.
}
\examples{
rankExpr(toy_expr) # toy_expr is a gene expression matrix, tiesMethod = 'min'
e <- Biobase::ExpressionSet(assayData = as.matrix(toy_expr))
rankExpr(e) # or it can be a ExpressionSet object
}
\seealso{
\code{\link{rank}} \linkS4class{GeneSet}
}
