% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{scoredf_ccle_ep}
\alias{scoredf_ccle_ep}
\title{Scoring results of a CCLE dataset against an epithelial gene signature}
\format{An object of class \code{data.frame} with 55 rows and 2 columns.}
\usage{
scoredf_ccle_ep
}
\description{
This data.frame is the returned results of function \code{\link[=singscoreing]{singscoreing()}} on a
CCLE dataset \href{https://www.nature.com/articles/nature11003}{Barretina et al}
consists of 55 samples agaist an epithelial gene signature obtained from
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4287932/}{Tan, Tuan Zea et al}
Please refer to the vignettes for the steps to aquaire the exact data object.
}
\references{
\insertRef{barretina2012cancer}{singscore}
\insertRef{Ep-MES}{singscore}
}
\seealso{
\link{scoredf_ccle_mes}
}
\keyword{datasets}
