% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{scoredf_tcga_epi}
\alias{scoredf_tcga_epi}
\title{Scoring results of TCGA tumour gene expression matrix}
\format{An object of class \code{data.frame} with 1119 rows and 2 columns.}
\usage{
scoredf_tcga_epi
}
\description{
This data.frame is the returned results of function \code{\link[=singscoring]{singscoring()}} on
tumour samples from \href{https://cancergenome.nih.gov}{TCGA} database against
an mesenchymal gene signature obtained from
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4287932/}{Tan, Tuan Zea et al}.
Please refer to the vignettes for the steps to aquaire the exact data object.
}
\references{
Foroutan, Momeneh, Joseph Cursons, Soroor Hediyeh-Zadeh,
Erik W Thompson, and Melissa J Davis. 2017.
“A Transcriptional Program for Detecting Tgfbeta-Induced Emt in Cancer.”
Molecular Cancer Research. American Association for Cancer Research.
doi:10.1158/1541-7786.MCR-16-0313.
}
\seealso{
\link{scoredf_tcga_mes}
}
\keyword{datasets}
