% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{scoredf_tumour_mes}
\alias{scoredf_tumour_mes}
\title{Scoring results of TCGA tumour gene expression matrix}
\format{An object of class \code{data.frame} with 1119 rows and 2 columns.}
\usage{
scoredf_tumour_mes
}
\description{
This data.frame is the returned results of function \code{\link[=singscoreing]{singscoreing()}} on
tumour samples from \href{https://cancergenome.nih.gov}{TCGA} database against
an epithelial gene signature obtained from
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4287932/}{Tan, Tuan Zea et al}.
Please refer to the
vignettes for the steps to aquaire the exact data object.
}
\references{
\insertRef{Ep-MES}{singscore}
}
\seealso{
\link{scoredf_tumour_ep}
}
\keyword{datasets}
