% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{package}
\name{singscore}
\alias{singscore}
\alias{singscore-package}
\title{singscore: A package for deriving gene-set scores at a single sample level.}
\description{
The package provides functions for calculating gene-set expression scores at
a single-sample level and it can deal with single-sample input. It also
includes efficient visualisation functions for exploring the scoring results
more in depth as well as functions for permutation test.
}
