% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankAndScoring.R
\name{singscoringOneGS}
\alias{singscoringOneGS}
\title{scoring single gene set signature when direction of gene set is unknown}
\usage{
singscoringOneGS(rankData, upSet, subSamples = NULL, centerScore = FALSE,
  dispersionFun = mad)
}
\arguments{
\item{rankData}{A matrix, ranked gene expression matrix data}

\item{upSet}{A GeneSet object, up regulated gene set}

\item{subSamples}{A character or vector of sample labels/indices that will be
used to subset the rankData matrix.All samples will be scored by default.}

\item{centerScore}{A Boolean, specifying whether scores should be centred,
default as FALSE}

\item{dispersionFun}{A character, dispersion function with default as 'mad'}

\item{downSet}{A GeneSet object, down regulated gene set}
}
\value{
A data.frame consists of scores and dispersions for all samples
}
\description{
scoring single gene set signature when direction of gene set is unknown
}
\keyword{internal}
