% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{tgfb_gs_dn}
\alias{tgfb_gs_dn}
\title{Down-set for TGFb gene expression signature}
\format{A GeneSet obeject}
\source{
\url{http://www.somelinktotgfbdata.info/}
}
\usage{
tgfb_gs_dn
}
\description{
A GeneSet object
}
\keyword{datasets}
