% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScoring.R
\docType{data}
\name{tgfb_gs_up}
\alias{tgfb_gs_up}
\title{Up-set for TGFb gene expression signature}
\format{A GeneSet obeject}
\source{
\url{http://www.somelinktotgfbdata.info/}
}
\usage{
tgfb_gs_up
}
\description{
A GeneSet object
}
\keyword{datasets}
