% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScoring.R
\docType{data}
\name{toy_dn}
\alias{toy_dn}
\title{Gene set of 5 genes}
\format{A GeneSet obeject with 5 genes}
\source{
\url{http://www.somelinktotgfbdata.info/}
}
\usage{
toy_dn
}
\description{
A GeneSet object that has ten genes
}
\keyword{datasets}
