% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{toy_gs_up}
\alias{toy_gs_up}
\title{A gene set object of up-regulated genes for the toy dataset}
\format{A GeneSet object with 5 genes}
\usage{
toy_gs_up
}
\description{
A GeneSet object with 5 genes randomly selected from the toy dataset. These
genes are independent of those in \link{toy_gs_dn}
}
\seealso{
\code{"\linkS4class{GeneSet}"},\link{toy_expr_se},\link{toy_gs_dn}
}
\keyword{datasets}
