% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{generateNull}
\alias{generateNull}
\title{Permutation test for the derived scores of each sample}
\usage{
generateNull(n_up, n_down, rankData, B = 1000, seed = 1)
}
\arguments{
\item{n_up}{integer, the number of genes in the up-regulated gene set}

\item{n_down}{integer, the number of genes in the down-regulated gene set}

\item{rankData}{matrix, outcome of function \code{\link[=rankGenes]{rankGenes()}}}

\item{B}{integer, the number of permutation repeats or the number of random
gene sets to be generated, default as 1000}

\item{seed}{integer, set the seed for randomisation}
}
\value{
A matrix of empirical scores for all samples
}
\description{
This function generates a number of random gene sets that
have the same number of genes as the scored gene set. It scores each random
gene set and returns a matrix of scores for all samples.
The empirical scores are used to calculate the empirical p-values and plot
the null distribution. The implementation uses \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}
for easy access to parallel backends.
}
\examples{
ranked <- rankGenes(toy_expr)
scoredf <- simpleScore(ranked, upSet = toy_gs_up, downSet = toy_gs_dn)
n_up = length(GSEABase::geneIds(toy_gs_up))
n_down = length(GSEABase::geneIds(toy_gs_dn))
# find out what backends can be registered on your machine
# the first one is the default backend, and it can be changed explicitly.
BiocParallel::registered()

# call the permutation function to generate the empirical scores for B times
permuteResult = generateNull(n_up = n_up, n_down = n_down, ranked, B =10,
seed = 1) 

}
\seealso{
\href{http://lcolladotor.github.io/2016/03/07/BiocParallel/#.WgXMF61L28U}{Post about BiocParallel}
\code{browseVignettes("BiocParallel")}
}
\author{
Ruqian Lyu
}
