% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{plotDispersion}
\alias{plotDispersion}
\title{Plot the score v.s. despersion for all samples}
\usage{
plotDispersion(scoredf, annot = NULL, alpha = 1, size = 1,
  textSize = 1.5, isInteractive = FALSE)
}
\arguments{
\item{scoredf}{data.frame, results of the singscoring() function}

\item{annot}{any annotation provided by the user that needs to be plot
annot must be ordered in the same was as the scores}

\item{alpha}{numeric, set the transparency of points}

\item{size}{numeric, Set the size of each point}

\item{textSize}{numeric, relative text sizes for title, labels and axis
values}

\item{isInteractive}{Boolean, Determine whether the plot is interactive}
}
\value{
A ggplot object
}
\description{
This function takes the output from the singscoring() function
and plots scatter plots for the score vs the dispersion for the total
score, the up score and the down score of samples. The plots
}
\examples{
ranked <- rankExpr(toy_expr)
scoredf <- singscoring(ranked, upSet = toy_up, downSet = toy_dn)
plotDispersion(scoredf)
plotDispersion(scoredf, isInteractive = TRUE)
}
