% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{plotNull}
\alias{plotNull}
\title{Plot the empirical null distribution using the permutation result}
\usage{
plotNull(permuResult, scoredf, pvals, sampleNames = NULL, cutoff = 0.01,
  alpha = 1, size = 1, textSize = 2, labelSize = 5)
}
\arguments{
\item{permuResult}{A matrix, outcome from function \code{\link[=permuteScores]{permuteScores()}}}

\item{scoredf}{A dataframe, outcome from function \code{\link[=singscoring]{singscoring()}}}

\item{pvals}{A vector, outcome of function \code{\link[=getPvals]{getPvals()}}}

\item{sampleNames}{A vector of character, sample names or multiple sample
labels}

\item{cutoff}{double, the cutoff value for determining significance}

\item{alpha}{numeric,ggplot theme element}

\item{size}{numeric,ggplot theme element}

\item{textSize}{numeric,ggplot theme element}

\item{labelSize}{numeric, size of label texts.}
}
\value{
a ggplot object
}
\description{
This function takes the results from function \code{\link[=permuteScores]{permuteScores()}}
and plots the density curves of empirical scores for the provided samples via
\code{sampleNames} parameter. It can plot null distribution for a single
sample or multiple samples.
}
\examples{
ranked <- rankExpr(toy_expr)
scoredf <- singscoring(ranked, upSet = toy_up, downSet = toy_dn)
n_up = length(GSEABase::geneIds(toy_up))
n_down = length(GSEABase::geneIds(toy_dn))
# find out what backends can be registered on your machine
BiocParallel::registered()
# the first one is the default backend, and it can be changed explicitly.
permuteResult = permuteScores(n_up = n_up, n_down = n_down, ranked, B =10,
seed = 1) # call the permutation function to generate the empirical scores 
#for B times.
pvals <- getPvals(permuteResult,scoredf)
plotNull(permuteResult,scoredf,pvals,sampleNames = names(pvals))
plotNull(permuteResult,scoredf,pvals,sampleNames = names(pvals)[1])
}
\author{
Ruqian Lyu
}
