% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNull}
\alias{plotNull}
\title{Plot the empirically estimated null distribution and associated p-values}
\usage{
plotNull(permuteResult, scoredf, pvals, sampleNames = NULL,
  cutoff = 0.01, textSize = 2, labelSize = 5)
}
\arguments{
\item{permuteResult}{A matrix, null distributions for each sample generated
using the \code{\link[=generateNull]{generateNull()}} function}

\item{scoredf}{A dataframe, singscores generated using the \code{\link[=simpleScore]{simpleScore()}}
function}

\item{pvals}{A vector, estimated p-values using the \code{\link[=getPvals]{getPvals()}} function
\code{permuteResult},\code{scoredf} and \code{pvals} are the results for the same samples.}

\item{sampleNames}{A character vector, sample IDs for which null
distributions will be plotted}

\item{cutoff}{numeric, the cutoff value for determining significance}

\item{textSize}{numeric, size of axes labels, axes values and title}

\item{labelSize}{numeric, size of label texts}
}
\value{
a ggplot object
}
\description{
This function takes the results from function \code{\link[=generateNull]{generateNull()}}
and plots the density curves of permuted scores for the provided samples via
\code{sampleNames} parameter. It can plot null distribution(s) for a single
sample or multiple samples.
}
\examples{
ranked <- rankGenes(toy_expr_se)
scoredf <- simpleScore(ranked, upSet = toy_gs_up, downSet = toy_gs_dn)
# find out what backends can be registered on your machine
BiocParallel::registered()
# the first one is the default backend, and it can be changed explicitly.
permuteResult = generateNull(upSet = toy_gs_up, downSet = toy_gs_dn, ranked, 
B =10, seed = 1,useBPPARAM = NULL) 
# call the permutation function to generate the empirical scores 
#for B times.
pvals <- getPvals(permuteResult,scoredf)
# plot for all samples
plotNull(permuteResult,scoredf,pvals,sampleNames = names(pvals))
#plot for the first sample
plotNull(permuteResult,scoredf,pvals,sampleNames = names(pvals)[1])
}
\author{
Ruqian Lyu
}
