% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{plotRankDensity}
\alias{plotRankDensity}
\title{Plot the density of ranks for one sample}
\usage{
plotRankDensity(rankData, upSet, downSet = NULL, isInteractive = F,
  textSize = 1.5)
}
\arguments{
\item{rankData}{one column of the data.frame obtained from the
rankExpr()/rankGenes() function}

\item{upSet}{GeneSet object, up regulated gene set}

\item{downSet}{GeneSet object, down regulated gene set}

\item{isInteractive}{Boolean, determin whether the returned plot is
interactive}

\item{textSize}{numberic, set the size of text on the plot}
}
\value{
A ggplot object (optionally interactive) demonstrating the rank
density along with rug plot
}
\description{
This function takes a single column data frame, which is a
subset of the ranked data obtained from rankGenes() function, and gene sets,
and it returns plots visualising the density and the rugs of the ranks.
}
\examples{
ranked <- rankExpr(toy_expr)

}
