% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{plotScoreLandscape}
\alias{plotScoreLandscape}
\title{Plot landscape of two gene signatures scores}
\usage{
plotScoreLandscape(scoredf1, scoredf2, scorenames = c(), textSize = 1.5,
  isInteractive = F, hexMin = 100)
}
\arguments{
\item{scoredf1}{data.frame, results of the simpleScore() or singscoring() function}

\item{scoredf2}{data.frame, results of the simpleScore() or singscoring() function}

\item{scorenames}{character, names for the two gene signatures scored stored
in scoredf1 and scoredf2}

\item{textSize}{numeric, relative text sizes for title, labels and axis
values}

\item{isInteractive}{boolean, whether the plot is interactive}
}
\value{
A ggplot object, a scatter plot, demostrating the relationship
between scores from two signatures on the same set of samples.
}
\description{
This function takes two data frames which are the output from
the simpleScore() function or singscoring() function and plots the relationship
between the two scores.
}
\examples{
ranked <- rankGenes(toy_expr)
scoredf <- simpleScore(ranked, upSet = toy_up, downSet = toy_dn)
scoredf2 <- simpleScore(ranked, upSet = toy_up)
plotScoreLandscape(scoredf, scoredf2)
}
