% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{rankExpr}
\alias{rankExpr}
\title{Rank gene expression matrix}
\usage{
rankExpr(exprsM, tiesMethod = "min")
}
\arguments{
\item{exprsM}{A matrix, gene expression matrix}

\item{tiesMethod}{A character, default as 'min'}
}
\value{
A matrix that has samples in colunm and genes in rows. Values are the
ranks of each gene in each sample.
}
\description{
Given a gene expression matrix and a tiesMethod (character),
this fucntion calls the 'rank' function from 'base' package which ranks the
gene expression matrix by gene's expression level. The default tiesMethod
is 'min'. There is a generic version of this function, details can be found
via the link in the see also section down at the bottom. It is suggested to
use the generic function 'rankExpr' which can accept multiple data formats
as input.
}
\examples{
\dontrun{ranked <- rankExpr(toy_expr)}
}
\seealso{
\code{\link[=rankExpr]{rankExpr()}}
}
\keyword{internal}
