% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{rankGenes}
\alias{rankGenes}
\title{Rank gene expression matrix}
\usage{
rankGenes(exprsM, tiesMethod = "min")
}
\arguments{
\item{exprsM}{A matrix, gene expression matrix}

\item{tiesMethod}{A character, default as 'min'}
}
\value{
A matrix that has samples in colunm and genes in rows. Values are the
  ranks of each gene in each sample.
}
\description{
Given a gene expression matrix and a tiesMethod
  character, this fucntion calls the 'rank' function in the base package
  The default ties.Method is set to 'min'. There is a generic version of
  this function, details can be found in the see also section down at the
  bottom. It is suggested to use the generic function 'rankExpr'
}
\examples{
ranked <- rankGenes(toy_expr)
}
\seealso{
\code{ \link{rankExpr}}
}
