% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankGenesGeneric.R
\docType{methods}
\name{rankGenes}
\alias{rankGenes}
\alias{rankGenes,matrix-method}
\alias{rankGenes,data.frame-method}
\alias{rankGenes,DGEList-method}
\alias{rankGenes,ExpressionSet-method}
\title{Rank genes by the gene expression intensities}
\usage{
rankGenes(expreMatrix, tiesMethod = "min")

\S4method{rankGenes}{matrix}(expreMatrix, tiesMethod = "min")

\S4method{rankGenes}{data.frame}(expreMatrix, tiesMethod = "min")

\S4method{rankGenes}{DGEList}(expreMatrix, tiesMethod = "min")

\S4method{rankGenes}{ExpressionSet}(expreMatrix, tiesMethod = "min")
}
\arguments{
\item{expreMatrix}{A gene expression matrix or S4 object
(matrix,data.frame,ExpressionSet,DGEList)}

\item{tiesMethod}{A character indicating what method to use when dealing with
ties}
}
\value{
The ranked gene expression matrix that has samples in columns and
genes in rows
}
\description{
The \code{rankGenes} function is a generic function that can deal
with mutilple types of inputs. Given a matrix of gene expression that has
samples in columns, genes in rows, and values being gene expression
intensity. \code{rankGenes} ranks gene expression in each sample. Given a
gene expression matrix (i.e matrix, data.frame) or a S4 object that has gene
expression matrix as a component (i.e ExpressionSet, DGEList) and a
'tiesMethod', it calls the  \code{rank} function in the base package which
ranks the gene expression matrix by its absolute expression level. If input
is S4 object of \code{DGEList, or ExpressionSet}, the generic function will
extract the gene epxression matrix genes on the rows and samples on columns
from the object and rank the genes. The default ties.Method is set to 'min'
}
\examples{
rankGenes(toy_expr) # toy_expr is a gene expression matrix, 
tiesMethod = 'min'
# or it can be a ExpressionSet object
e <- Biobase::ExpressionSet(assayData = as.matrix(toy_expr))
rankGenes(e) 
}
\seealso{
\code{\link{rank}} \link[=ExpressionSet-class]{ExpressionSet}
\link[=DGEList-class]{DGEList}
}
