% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{scoredf_ccle_mes}
\alias{scoredf_ccle_mes}
\title{Pre-computed singscores of the CCLE dataset against a mesenchymal gene
signature}
\format{An object of class \code{data.frame} with 55 rows and 2 columns.}
\usage{
scoredf_ccle_mes
}
\description{
This data.frame stores pre-computed scores of the CCLE dataset \href{https://www.nature.com/articles/nature11003}{Barretina etal} calculated using the
\code{\link[=simpleScore]{simpleScore()}} function against the mesenchymal gene signature from \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4287932/}{Tan,Tuan Zea et al}. The
data.frame has scores for 55 samples. Please refer to the vignettes for
instructions on how to obtain the full datasets.
}
\references{
Barretina, Jordi, Giordano Caponigro, Nicolas Stransky, Kavitha
Venkatesan, Adam A Margolin, Sungjoon Kim, Christopher J Wilson, et al. 2012.
“The Cancer Cell Line Encyclopedia Enables Predictive Modelling of Anticancer
Drug Sensitivity.” Nature 483 (7391): 603–7.

Tan, Tuan Zea, Qing Hao Miow, Yoshio Miki, Tetsuo Noda, Seiichi Mori, Ruby
Yun-Ju Huang, and Jean Paul Thiery. 2014–10AD. “Epithelial-Mesenchymal
Transition Spectrum Quantification and Its Efficacy in Deciphering Survival
and Drug Responses of Cancer Patients.” EMBO Molecular Medicine 6 (10).
Oxford, UK: BlackWell Publishing Ltd: 1279–93 doi:10.15252/emmm.201404208.
}
\seealso{
\code{[scoredf_ccle_epi]}
}
\keyword{datasets}
