% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{scoredf_tcga_epi}
\alias{scoredf_tcga_epi}
\title{Pre-computed scores of the TCGA breast cancer gene expression matrix
against an epithelial signature}
\format{
An object of class \code{data.frame} with 1119 rows and 2 columns.
}
\usage{
scoredf_tcga_epi
}
\description{
This data.frame stores pre-computed scores of the
\href{https://cancergenome.nih.gov}{TCGA} dataset calculated using the
\code{\link[=simpleScore]{simpleScore()}} function against the epithelial gene signature from \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4287932/}{Tan, Tuan Zea et al}.
Please refer to the vignettes for instructions on how to obtain the full
datasets.
}
\references{
Tan, Tuan Zea, Qing Hao Miow, Yoshio Miki, Tetsuo Noda, Seiichi
Mori, Ruby Yun-Ju Huang, and Jean Paul Thiery. 2014–10AD.
“Epithelial-Mesenchymal Transition Spectrum Quantification and Its Efficacy
in Deciphering Survival and Drug Responses of Cancer Patients.” EMBO
Molecular Medicine 6 (10). Oxford, UK: BlackWell Publishing Ltd: 1279–93
doi:10.15252/emmm.201404208.
}
\seealso{
\link{scoredf_tcga_mes}
}
\keyword{datasets}
