% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScoreGeneric.R
\docType{methods}
\name{simpleScore}
\alias{simpleScore}
\alias{simpleScore,ANY,vector,missing-method}
\alias{simpleScore,ANY,GeneSet,missing-method}
\alias{simpleScore,ANY,vector,vector-method}
\alias{simpleScore,ANY,GeneSet,GeneSet-method}
\alias{simpleScore,ANY,GeneSet,vector-method}
\alias{simpleScore,ANY,vector,GeneSet-method}
\title{single-sample gene-set scoring method}
\usage{
simpleScore(rankData, upSet, downSet = NULL, subSamples = NULL,
  centerScore = TRUE, dispersionFun = "mad")

\S4method{simpleScore}{ANY,vector,missing}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{simpleScore}{ANY,GeneSet,missing}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{simpleScore}{ANY,vector,vector}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{simpleScore}{ANY,GeneSet,GeneSet}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{simpleScore}{ANY,GeneSet,vector}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{simpleScore}{ANY,vector,GeneSet}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")
}
\arguments{
\item{rankData}{A matrix-like object, ranked gene expression matrix data}

\item{upSet}{A GeneSet object or vector of gene Ids of up-regulated gene set}

\item{downSet}{A GeneSet object or vector of gene Ids of down-regulated gene
set}

\item{subSamples}{A vector of sample labels/indices that will be
used to subset the rankData matrix. All samples will be scored by default.}

\item{centerScore}{A Boolean, specifying whether scores should be centered,
default as TRUE}

\item{dispersionFun}{A character, dispersion function with default as 'mad'}
}
\value{
A data.frame consists of scores and dispersions for all samples
}
\description{
This function takes a ranked gene expression matrix obtained from
\code{rankGenes()} function and gene sets as inputs and it calculates the
scores for each individual sample against gene set. It returns a data.frame
consists of scores and dispersions for each sample. The gene sets can be in
vector format or GeneSet S4 object (GSEABase package).
Down set can be null if down-regulated gene sets are unavailable.
}
\examples{
ranked <- rankGenes(toy_expr)
scoredf <- simpleScore(ranked, upSet = toy_gs_up, downSet = toy_gs_dn)
# toy_gs_up is a GeneSet object, alternatively a vector of gene ids may also 
# be supplied.
}
\seealso{
\code{\link{rank}}
\code{"\linkS4class{GeneSet}"}
}
