% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScoreGeneric.R
\name{simpleScore}
\alias{simpleScore}
\alias{simpleScore,matrix,vector,missing-method}
\alias{simpleScore,matrix,GeneSet,missing-method}
\alias{simpleScore,matrix,vector,vector-method}
\alias{simpleScore,matrix,GeneSet,GeneSet-method}
\title{single-sample gene-set scoring method}
\usage{
simpleScore(
  rankData,
  upSet,
  downSet = NULL,
  subSamples = NULL,
  centerScore = TRUE,
  dispersionFun = mad,
  knownDirection = TRUE
)

\S4method{simpleScore}{matrix,vector,missing}(
  rankData,
  upSet,
  downSet = NULL,
  subSamples = NULL,
  centerScore = TRUE,
  dispersionFun = mad,
  knownDirection = TRUE
)

\S4method{simpleScore}{matrix,GeneSet,missing}(
  rankData,
  upSet,
  downSet = NULL,
  subSamples = NULL,
  centerScore = TRUE,
  dispersionFun = mad,
  knownDirection = TRUE
)

\S4method{simpleScore}{matrix,vector,vector}(
  rankData,
  upSet,
  downSet = NULL,
  subSamples = NULL,
  centerScore = TRUE,
  dispersionFun = mad,
  knownDirection = TRUE
)

\S4method{simpleScore}{matrix,GeneSet,GeneSet}(
  rankData,
  upSet,
  downSet = NULL,
  subSamples = NULL,
  centerScore = TRUE,
  dispersionFun = mad,
  knownDirection = TRUE
)
}
\arguments{
\item{rankData}{A matrix object, ranked gene expression matrix data generated
using the \code{\link[=rankGenes]{rankGenes()}} function (make sure this matrix is not modified)}

\item{upSet}{A GeneSet object or character vector of gene IDs of up-regulated
gene set or a gene set where the nature of genes is not known}

\item{downSet}{A GeneSet object or character vector of gene IDs of
down-regulated gene set or NULL where only a single gene set is provided}

\item{subSamples}{A vector of sample labels/indices that will be used to
subset the rankData matrix. All samples will be scored if not provided}

\item{centerScore}{A Boolean, specifying whether scores should be centered
around 0, default as TRUE. Note: scores never centered if \code{knownDirection = FALSE}}

\item{dispersionFun}{A function, dispersion function with default being \code{mad}}

\item{knownDirection}{A boolean, determining whether the gene set should be
considered to be directional or not. A gene set is directional if the type
of genes in it are known i.e. up- or down-regulated. This should be set to
TRUE if the gene set is composed of both up- AND down-regulated genes.
Defaults to TRUE. This parameter becomes irrelevant when both upSet(Colc)
and downSet(Colc) are provided.}
}
\value{
A data.frame consists of singscores and dispersions for all samples
}
\description{
This function computes 'singscores' using an \strong{unmodified}ranked
gene expression matrix obtained from the \code{\link[=rankGenes]{rankGenes()}} function and a gene
set or a pair of up-regulated and down-regulated gene sets. It returns a
data.frame of scores and dispersions for each sample. The gene sets can be
in vector format or as GeneSet objects (from GSEABase packages). If samples
need to be scored against a single gene set, the \code{upSet} argument
should be used to pass the gene set while the \code{downSet} argument is set
to \code{NULL}. This setting is ideal for gene sets representing gene
ontologies where the nature of the genes is unknown (up- or down-regulated).
}
\examples{
ranked <- rankGenes(toy_expr_se)
scoredf <- simpleScore(ranked, upSet = toy_gs_up, downSet = toy_gs_dn)
# toy_gs_up is a GeneSet object, alternatively a vector of gene ids may also
# be supplied.
}
\seealso{
\code{\link{rank}} \code{"\linkS4class{GeneSet}"}
}
