% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{simpleScore}
\alias{simpleScore}
\title{Calculate scores for the ranked gene expression matrix against a gene set}
\usage{
simpleScore(rankData, subSamples = NULL, upSet, downSet = NULL,
  centerScore = TRUE, dispersionFun = mad)
}
\arguments{
\item{rankData}{A matrix, ranked gene expression matrix data}

\item{subSamples}{A character or vector of sample labels/indices that will be
used to subset the rankData matrix}

\item{upSet}{A GeneSet object, up regulated gene set}

\item{downSet}{A GeneSet object, down regulated gene set}

\item{centerScore}{A Boolean, specifying whether scores should be centred}

\item{dispersionFun}{A character, dispersion function with default as 'mad'}
}
\value{
A data.frame consists of scores and dispersions for all samples
}
\description{
This function takes a ranked gene expression matrix obtained
from \code{rankExpr} (or \code{rankGenes}) function and a GeneSet object as
input parameters. It returns a data.frame consists of scores and
dispersions for each sample.
}
\examples{
ranked <- rankGenes(toy_expr)
scoredf <- simpleScore(ranked, upSet = toy_up, downSet = toy_dn)
}
\seealso{
\link[=GeneSet-class]{GeneSet}
\code{\link[=singscoring]{singscoring()}}
}
