% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFuns.R
\name{simpleScore}
\alias{simpleScore}
\title{Single-sample Gene-set scoring method}
\usage{
simpleScore(rankData, upSet, downSet = NULL, subSamples = NULL,
  centerScore = TRUE, dispersionFun = mad)
}
\arguments{
\item{rankData}{A matrix, ranked gene expression matrix data}

\item{upSet}{A GeneSet object, up regulated gene set}

\item{downSet}{A GeneSet object, down regulated gene set}

\item{subSamples}{A character or vector of sample labels/indices that will be
used to subset the rankData matrix.All samples will be scored by default.}

\item{centerScore}{A Boolean, specifying whether scores should be centred}

\item{dispersionFun}{A character, dispersion function with default as 'mad'}
}
\value{
A data.frame consists of scores and dispersions for all samples
}
\description{
This function takes a ranked gene expression matrix obtained
from \code{rankExpr}  function and a GeneSet object as input parameters. It
returns a data.frame consists of calculated scores and dispersions for each
sample. It is suggested to use the generic version of of this method which
can work with gene set stored in both vector or GeneSet.
}
\examples{
ranked <- rankExpr(toy_expr)
scoredf <- simpleScore(ranked, upSet = toy_up, downSet = toy_dn)
}
\seealso{
\code{\link[=singscoring]{singscoring()}}
\link[=GeneSet-class]{GeneSet}
}
