% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{package}
\name{singscore}
\alias{singscore}
\title{singscore: A package for deriving gene-set scores at a single sample level}
\description{
The package provides functions for calculating gene-set enrichment scores at
a single-sample level using gene expression data. It includes functions to
perform hypothesis testing and provides visualisations to enable diagnosis of
scores and gene sets along with visualisations to enable exploration of
results.
}
