% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{package}
\name{singscore}
\alias{singscore}
\alias{singscore-package}
\title{singscore: A package for deriving gene-set score at a single sample level.}
\description{
The packge provides functions of calculating gene-set expression scores at a
single-sample level and it is able to deal with single-sample input. It also
includes efficient visulisation function for exploring the results as well
as permutation test for significance test and plot null distribution.
}
