% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscoringGeneric.R
\docType{methods}
\name{singscoring}
\alias{singscoring}
\alias{singscoring,ANY,vector,missing-method}
\alias{singscoring,ANY,GeneSet,missing-method}
\alias{singscoring,ANY,vector,vector-method}
\alias{singscoring,ANY,GeneSet,GeneSet-method}
\alias{singscoring,ANY,GeneSet,vector-method}
\alias{singscoring,ANY,vector,GeneSet-method}
\title{single-sample gene-set scoring method}
\usage{
singscoring(rankData, upSet, downSet = NULL, subSamples = NULL,
  centerScore = TRUE, dispersionFun = "mad")

\S4method{singscoring}{ANY,vector,missing}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{singscoring}{ANY,GeneSet,missing}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{singscoring}{ANY,vector,vector}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{singscoring}{ANY,GeneSet,GeneSet}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{singscoring}{ANY,GeneSet,vector}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")

\S4method{singscoring}{ANY,vector,GeneSet}(rankData, upSet, downSet = NULL,
  subSamples = NULL, centerScore = TRUE, dispersionFun = "mad")
}
\arguments{
\item{rankData}{A matrix-like object, ranked gene expression matrix data}

\item{upSet}{A GeneSet object or vector of gene ids of up regulated gene set}

\item{downSet}{A GeneSet object or vector of gene ids of down regulated gene
set}

\item{subSamples}{A vector of sample labels/indices that will be
used to subset the rankData matrix. All samples will be scored by default.}

\item{centerScore}{A Boolean, specifying whether scores should be centred,
default as TRUE}

\item{dispersionFun}{A character, dispersion function with default as 'mad'}
}
\value{
A data.frame consists of scores and dispersions for all samples
}
\description{
This function takes a ranked gene expression matrix obtained from
\code{rankExpr} (or \code{rankGenes}) function and gene sets as input and
calulate the scores for each individual sample against gene set. It returns
a data.frame consists of scores and dispersions for each sample. The gene
sets can be in vector format or GeneSet S4 object. Down set can be null if
no down-regulated gene sets are available.
}
\examples{
ranked <- rankExpr(toy_expr)
scoredf <- singscoring(ranked, upSet = toy_up, downSet = toy_dn)
# toy_up is a GeneSet object, alternatively a vector of gene ids may also 
# be supplied.
}
\seealso{
\code{\link{rank}}
\code{"\linkS4class{GeneSet}"}
}
