% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscoringGeneric.R
\docType{methods}
\name{singscoring}
\alias{singscoring}
\alias{singscoring,matrix,ANY,vector,vector-method}
\title{Calculate scores for the ranked gene expression matrix against gene set}
\usage{
singscoring(rankData, subSamples, upSet, downSet, centerScore, dispersionFun)

\S4method{singscoring}{matrix,ANY,vector,vector}(rankData, subSamples, upSet,
  downSet, centerScore, dispersionFun)
}
\arguments{
\item{rankData}{A matrix-like object, ranked gene expression matrix data}

\item{subSamples}{A character or vector of sample labels/indices that will be
used to subset the rankData matrix}

\item{upSet}{A GeneSet object or vector of gene ids, up regulated gene set}

\item{downSet}{A GeneSet object or vector of gene ids, down regulated gene set}

\item{centerScore}{A Boolean, specifying whether scores should be centred}

\item{dispersionFun}{A character, dispersion function with default as 'mad'}
}
\value{
A data.frame consists of scores and dispersions for all samples
}
\description{
This function takes a ranked gene expression matrix obtained
  from \code{rankExpr} (or \code{rankGenes}) function and gene sets as input.
  It returns a data.frame consists of scores and dispersions for each sample.
  The gene sets can be in vector format or GeneSet S4 object.
}
\examples{
ranked <- rankExpr(toy_expr)
scoredf <- singscoring(ranked, upSet = toy_up, downSet = toy_dn)

}
\seealso{
\code{\link{rank}} \linkS4class{GeneSet}
}
