% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{tgfb_expr_10}
\alias{tgfb_expr_10}
\title{Gene expression data.frame}
\format{A data.frame obeject}
\source{
\url{http://www.somelinktotgfbdata.info/}
}
\usage{
tgfb_expr_10
}
\description{
A data.frame object that has 10 samples selected from a TFGb full dataset
}
\keyword{datasets}
