% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{toy_dn}
\alias{toy_dn}
\title{Gene set object}
\format{A GSEABase::GeneSet obeject with 5 genes}
\usage{
toy_dn
}
\description{
A GeneSet object that represents the down-regulated gene set in a gene
signature and it consists of five genes which are randomly selected from the
gene list in \code{toy_expr} but has no overlap with \code{toy_up}.
}
\seealso{
\link[=GeneSet-class]{GeneSet},\link{toy_expr},\link{toy_up}
}
\keyword{datasets}
