% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScoring.R
\docType{data}
\name{toy_expr}
\alias{toy_expr}
\title{Gene expression dataset of 100 genes in ten samples}
\format{A data frame with 100 rows and 10 columns:
\describe{
\item{D_Ctrl_R1}{a control sample label}
\item{D_TGFb_R1}{a tgfb sample label}
 ...
 }}
\source{
\url{http://www.somelinktotgfbdata.info/}
}
\usage{
toy_expr
}
\description{
A gene by sample matrix-like dataset with value representing the gene's
expression intensity in a sample
}
\keyword{datasets}
