% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\docType{data}
\name{toy_gs_up}
\alias{toy_gs_up}
\title{Gene set of 5 genes}
\format{A GeneSet obeject with 5 genes}
\usage{
toy_gs_up
}
\description{
A GeneSet object that represents the up-regulated gene set in a gene
signature and it consists of five genes which are randomly selected from the
gene list in \code{toy_expr} but has no overlap with \code{toy_gs_dn}.
}
\seealso{
\code{"\linkS4class{GeneSet}"},\link{toy_expr},\link{toy_gs_dn}
}
\keyword{datasets}
